consolidate compat feature impls/tests/benchs to compat module
